Changed interpolate_parameters to use xarray
Also added a jupyter notebook to
* merge parameters into one netcdf file
* write parameters from single merged netcdf file back to single pickle
  files, one per grid cell. After that they have the same form as output
by pymc
* test that merging parameters and writing them back to single files
  does not change anything

The first function is already implemented in interpolate_parameters.py
in this commit. The second function still needs to be included into the
main code. The third function may be implemented as unittest. When all
functions are implemented the notebook may be deleted